refactor(a11y): server-render titles for proposal/post pages
Convert the proposal-view, post, and legacy decision/proposal pages from
client-side document.title to server generateMetadata + prefetch/hydrate.

- Title data is fetched in generateMetadata and the page prefetch through
  the shared cache()-wrapped createServerUtils with a staleTime, so each
  query (including the ones that fire a "viewed" analytics event) resolves
  once per request — no double-count, real server-rendered <title>
- Fixes the legacy decision page double-firing trackProcessViewed
- Client pages now receive route params as props and hydrate from the
  dehydrated query cache instead of refetching in the browser
- The collaborative proposal editor stays on client document.title by
  design (noindex, JS-required, soft-nav only; conversion risk not worth it)
